Rename membership app as member
