+++ asso/member/constants.py
