+++ asso/member/fixtures/member/demo.yaml
@@ -1,40 +1,40 @@
-- model: membership.MemberQualification
+- model: member.MemberQualification
   pk: 1
   fields:
     slug: master
     title: "Master Universitario"
     order: 10
-- model: membership.MemberQualification
+- model: member.MemberQualification
   pk: 2
   fields:
     slug: phd
     title: "Dottorato di Ricerca"
     order: 20
-- model: membership.MemberQualification
+- model: member.MemberQualification
   pk: 3
   fields:
     slug: master-degree
     title: "Laurea Magistrale"
     order: 30
-- model: membership.MemberQualification
+- model: member.MemberQualification
   pk: 4
   fields:
     slug: bachelor-degree
     title: "Laurea Triennale"
     order: 40
-- model: membership.MemberQualification
+- model: member.MemberQualification
   pk: 5
   fields:
     slug: high-school
     title: "Diploma Maturità"
     order: 50
-- model: membership.MemberQualification
+- model: member.MemberQualification
   pk: 6
   fields:
     slug: mid-school
     title: "Licenza Media"
     order: 60
-- model: membership.MemberQualification
+- model: member.MemberQualification
   pk: 7
   fields:
     slug: primary-school
