+++ asso/member/views.py
@@ -4,7 +4,7 @@ from . import models
 
 
 class MemberDetailView(generic.DetailView):
-    template_name = "membership/member.html"
+    template_name = "member/member.html"
     context_object_name = "member"
     model = models.Member
